Refactor timeline and Actual client internals
